requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
factory_boy==3.3.1
Faker==33.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.0.0
msgpack==1.1.0
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.0
service-identity==24.2.0
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
zope.interface==7.2
